requirements.txt
aiosqlite==0.16.1
appdirs==1.4.4
asgiref==3.4.1
asyncpg==0.23.0
black==21.6b0
click==8.0.1
colorama==0.4.4
fastapi==0.65.2
h11==0.12.0
iso8601==0.1.14
isort==5.9.1
mypy-extensions==0.4.3
pathspec==0.8.1
pydantic==1.8.2
pypika-tortoise==0.1.1
pytz==2021.1
regex==2021.7.1
starlette==0.14.2
toml==0.10.2
tortoise-orm==0.17.4
typing-extensions==3.10.0.0
uvicorn==0.14.0
